chore(alert): move alert and user entities to infrastructure
Keep legacy package typealias shims for compatibility while the real JPA entities live under infrastructure.adapter.output. Preserve entity graphs, relationships, and lifecycle validation.
